Move ipv8 into component folder
