+++ experiment/tunnel_community/speed_test_exit.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from ipv8.messaging.anonymization.tunnel import EXIT_NODE, ORIGINATOR
 from ipv8.messaging.anonymization.utils import run_speed_test
 from ipv8.taskmanager import TaskManager
-from tribler_core.components.ipv8 import Ipv8Component
+from tribler_core.components.ipv8.ipv8_component import Ipv8Component
 from tribler_core.components.masterkey.masterkey_component import MasterKeyComponent
 from tribler_core.components.restapi import RESTComponent
 from tribler_core.components.tunnels import TunnelsComponent
