+++ src/tribler-core/tribler_core/components/ipv8/ipv8_component.py
