+++ src/tribler-core/tribler_core/components/payout.py
@@ -1,7 +1,7 @@
 from tribler_common.simpledefs import NTFY
 from tribler_core.components.bandwidth_accounting.bandwidth_accounting_component import BandwidthAccountingComponent
 from tribler_core.components.base import Component
-from tribler_core.components.ipv8 import Ipv8Component
+from tribler_core.components.ipv8.ipv8_component import Ipv8Component
 from tribler_core.components.reporter import ReporterComponent
 from tribler_core.modules.payout.payout_manager import PayoutManager
