+++ src/tribler-core/tribler_core/components/popularity.py
@@ -1,7 +1,7 @@
 from ipv8.peerdiscovery.network import Network
 from tribler_core.components.base import Component
 from tribler_core.components.gigachannel.community.sync_strategy import RemovePeers
-from tribler_core.components.ipv8 import INFINITE, Ipv8Component
+from tribler_core.components.ipv8.ipv8_component import INFINITE, Ipv8Component
 from tribler_core.components.metadata_store.metadata_store_component import MetadataStoreComponent
 from tribler_core.components.reporter import ReporterComponent
 from tribler_core.components.torrent_checker import TorrentCheckerComponent
